refactor: Move Recipient related components in a Recipient folder
Recipient file contained ~10 different components.
I moved them in a file per component to improve lisibility.
I also moved Recipient related component from the components folder
to the Recipient.

BREAKING CHANGE: If you imported component directly from
'cozy-sharing/dist/components/... '
especially from ' 'cozy-sharing/dist/components/Recipient',
you will need to update your import.

Drive example (the only one I found) :
`import { RecipientsAvatars } from 'cozy-sharing/dist/components/Recipient'`
=>
`import RecipientsAvatars from 'cozy-sharing/dist/components/Recipient/RecipientsAvatars'`
