+++ packages/cozy-sharing/src/SharingOwnerAvatar.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import SharingContext from './context'
 import withLocales from './withLocales'
-import { RecipientAvatar } from './components/Recipient'
+import RecipientAvatar from './components/Recipient/RecipientAvatar'
 
 export const SharingOwnerAvatar = withLocales(({ docId, ...rest }) => (
   <SharingContext.Consumer>
