+++ packages/cozy-sharing/src/components/Recipient/Identity.jsx
