+++ packages/cozy-sharing/src/components/Recipient/Recipient.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Recipient, { excludeMeAsOwnerFromRecipients } from './Recipient'
-import AppLike from '../../test/AppLike'
+import Recipient from './Recipient'
+import AppLike from '../../../test/AppLike'
 import { createMockClient } from 'cozy-client'
 import { render, fireEvent } from '@testing-library/react'
 
@@ -154,43 +154,5 @@ describe('Recipient component', () => {
     expect(verifyRecipient).toBeCalledWith({
       email: 'me@bob.cozy.localhost'
     })
-  })
-})
-
-describe('excludeMeAsOwnerFromRecipients', () => {
-  test('excludeMeAsOwnerFromRecipients behavior', () => {
-    const recipients = [
-      {
-        status: 'owner',
-        instance: 'http://foo1.cozy.bar'
-      },
-      { status: 'pending', instance: 'http://foo2.cozy.bar' },
-      {
-        status: 'pending',
-        instance: 'http://foo3.cozy.bar'
-      }
-    ]
-    const client = { options: { uri: 'http://foo2.cozy.bar' } }
-    expect(
-      excludeMeAsOwnerFromRecipients({ recipients, isOwner: true, client })
-    ).toEqual([
-      { status: 'pending', instance: 'http://foo2.cozy.bar' },
-      {
-        status: 'pending',
-        instance: 'http://foo3.cozy.bar'
-      }
-    ])
-    expect(
-      excludeMeAsOwnerFromRecipients({ recipients, isOwner: false, client })
-    ).toEqual([
-      {
-        status: 'owner',
-        instance: 'http://foo1.cozy.bar'
-      },
-      {
-        status: 'pending',
-        instance: 'http://foo3.cozy.bar'
-      }
-    ])
   })
 })
