+++ packages/cozy-sharing/src/components/Recipient/RecipientsAvatars.spec.jsx
@@ -2,8 +2,11 @@ import React from 'react'
 import { createMockClient } from 'cozy-client'
 import { render } from '@testing-library/react'
 
-import { RecipientsAvatars, MAX_DISPLAYED_RECIPIENTS } from './Recipient'
-import AppLike from '../../test/AppLike'
+import RecipientsAvatars, {
+  MAX_DISPLAYED_RECIPIENTS,
+  excludeMeAsOwnerFromRecipients
+} from './RecipientsAvatars'
+import AppLike from '../../../test/AppLike'
 
 const mockRecipients = new Array(MAX_DISPLAYED_RECIPIENTS - 1)
   .fill(
@@ -111,5 +114,43 @@ describe('RecipientsAvatars', () => {
 
     expect(getByTestId('recipientsAvatars-plusX')).toBeTruthy()
     expect(getByTestId('recipientsAvatars-link')).toBeTruthy()
+  })
+})
+
+describe('excludeMeAsOwnerFromRecipients', () => {
+  test('excludeMeAsOwnerFromRecipients behavior', () => {
+    const recipients = [
+      {
+        status: 'owner',
+        instance: 'http://foo1.cozy.bar'
+      },
+      { status: 'pending', instance: 'http://foo2.cozy.bar' },
+      {
+        status: 'pending',
+        instance: 'http://foo3.cozy.bar'
+      }
+    ]
+    const client = { options: { uri: 'http://foo2.cozy.bar' } }
+    expect(
+      excludeMeAsOwnerFromRecipients({ recipients, isOwner: true, client })
+    ).toEqual([
+      { status: 'pending', instance: 'http://foo2.cozy.bar' },
+      {
+        status: 'pending',
+        instance: 'http://foo3.cozy.bar'
+      }
+    ])
+    expect(
+      excludeMeAsOwnerFromRecipients({ recipients, isOwner: false, client })
+    ).toEqual([
+      {
+        status: 'owner',
+        instance: 'http://foo1.cozy.bar'
+      },
+      {
+        status: 'pending',
+        instance: 'http://foo3.cozy.bar'
+      }
+    ])
   })
 })
