+++ packages/cozy-sharing/src/components/Recipient/recipient.styl
